backend/requirements.txt
django>=4.0,<5.0
djangorestframework
django-cors-headers
gunicorn
whitenoise
python-dotenv
pillow
opencv-python-headless
numpy
tensorflow-cpu
google-genai
